+++ src/cpu/mos6502/instr/inc.rs
@@ -3,7 +3,7 @@ use cpu::mos6502::{ExecError,Mos6502,Operand};
 
 pub fn exec<M>(cpu: &mut Mos6502<M>, op: Operand) -> Result<(), ExecError> where M: Memory {
     let new_val = (try!(op.get_u8(cpu)) + 1) & 0xFF;
-    cpu.flags.set_sign_and_zero(new_val as usize);
+    cpu.flags.set_sign_and_zero(new_val);
     try!(op.set_u8(cpu, new_val));
     Ok(())
 }
