+++ src/cpu/mos6502/instr/pull.rs
@@ -3,7 +3,7 @@ use cpu::mos6502::{ExecError,Mos6502,RegisterName};
 
 pub fn exec<M>(cpu: &mut Mos6502<M>, r: RegisterName) -> Result<(), ExecError> where M : Memory {
     let val = try!(cpu.pop());
-    cpu.flags.set_sign_and_zero(val as usize);
+    cpu.flags.set_sign_and_zero(val);
     r.set(cpu, val);
     Ok(())
 }
